Make error reporting compatible with PHPUnit 5 and PHP 7
In working on the unit testing tutorial, I discovered two issues with how
zend-test reports caught exceptions.

`AbstractControllerTestCase::tearDown()` was testing for an exception in the
`MvcEvent`, and, if found, re-throwing it. Unfortunately, this breaks for two
reasons:

- In PHP 7, we need to test for `Throwable`, as zend-mvc v2 and v3 now will
  catch any `Throwable`.
- In PHPUnit 5, exceptions thrown in `tearDown()` are caught and silently
  discarded. Even passing the verbosity flag during execution will not display
  their messages.

I found the solution was to create a new method, `createFailureMessage()`, which
takes the base message to use as an argument. If no exception is present in the
`MvcEvent` *or* `$traceError` is disabled, we return the message unchanged.
Otherwise, we loop through the exception stack, gathering information (the
exception type, the exception message, and the file and line in which the
exception was thrown), creating a formatted message to append to the provided
message.

Additionally, when testing for an exception, we test against both `Throwable`
and `Exception`, to allow usage with either PHP 7 or PHP 5.6.

Interestingly, we've never had any tests for the `traceError` functionality
previously. The tests in this patch are not comprehensive, but demonstrate that
messages passed through the new message do get the exception details.