Added PHPUnit 8 compatibility level
Added traits with methods which can be used in tests and in PHPUnit 8 have
RTH defined. We load proper trait depends on used PHPUnit version.
In our classes to keep compatibility we use method with "Compat" suffix.
Fot the end user nothing is changed as long as the user decide to support
one PHPUnit version (any <7 or 8+). Depends on the version user can define
setUp/tearDown method compatible with the used version of PHPUnit and call
the parent method.